+++ modules/platforms/dotnet/Apache.Ignite.Tests/Sql/SqlTests.cs
@@ -23,19 +23,13 @@ namespace Apache.Ignite.Tests.Sql
     using System.Threading.Tasks;
     using Ignite.Sql;
     using Ignite.Table;
-    using NodaTime;
     using NUnit.Framework;
-    using Table;
 
     /// <summary>
     /// Tests for SQL API: <see cref="ISql"/>.
     /// </summary>
     public partial class SqlTests : IgniteTestsBase
     {
-        private const string AllColumnsQuery = "select \"KEY\", \"STR\", \"INT8\", \"INT16\", \"INT32\", \"INT64\", \"FLOAT\", " +
-                                               "\"DOUBLE\", \"DATE\", \"TIME\", \"DATETIME\", \"TIMESTAMP\", \"BLOB\", \"DECIMAL\" " +
-                                               "from TBL_ALL_COLUMNS_SQL ORDER BY KEY";
-
         [OneTimeSetUp]
         public async Task CreateTestTable()
         {
@@ -57,38 +51,6 @@ namespace Apache.Ignite.Tests.Sql
         {
             await Client.Sql.ExecuteAsync(null, "DROP TABLE TEST");
             await Client.Sql.ExecuteAsync(null, "DROP TABLE IF EXISTS TestDdlDml");
-        }
-
-        [OneTimeSetUp]
-        public async Task InsertTestData()
-        {
-            var pocoAllColumns1 = new PocoAllColumnsSqlNullable(
-                Key: 1,
-                Str: "v-1",
-                Int8: 2,
-                Int16: 3,
-                Int32: 4,
-                Int64: 5,
-                Float: 6.5F,
-                Double: 7.5D,
-                Date: new LocalDate(2023, 01, 18),
-                Time: new LocalTime(09, 28),
-                DateTime: new LocalDateTime(2023, 01, 18, 09, 29),
-                Timestamp: Instant.FromUnixTimeSeconds(123),
-                Blob: new byte[] { 1, 2 },
-                Decimal: 8.7M);
-
-            var pocoAllColumns2 = new PocoAllColumnsSqlNullable(
-                Key: 2,
-                Str: "v-2",
-                Int8: sbyte.MinValue,
-                Int16: short.MinValue,
-                Int32: int.MinValue,
-                Int64: long.MinValue,
-                Float: float.MinValue,
-                Double: double.MinValue);
-
-            await PocoAllColumnsSqlNullableView.UpsertAllAsync(null, new[] { pocoAllColumns1, pocoAllColumns2 });
         }
 
         [Test]
